chore: move ContextOptions to context.options
